backend/src/orm/database.py
from sqlalchemy.orm import sessionmaker
from sqlalchemy.ext.asyncio import create_async_engine, AsyncSession

URL = "sqlite+aiosqlite:///database.db"

async_engine = create_async_engine(URL, echo=False)

async_session = sessionmaker(
    bind=async_engine,
    class_=AsyncSession,
    expire_on_commit=False
)